Added normal parameter
Just to handle things right

+++ UrfTweaks.ps1
@@ -1,4 +1,5 @@
 param (
+    [Alias("n")][switch]$normal,
     [Alias("f")][switch]$full,
     [Alias("s")][switch]$silent,
     [Alias("r")][switch]$restart
@@ -25,22 +26,22 @@ Import-Module -DisableNameChecking "$PSScriptRoot\modules\titles.psm1"
 
 Clear-Host
 if ($full) { Write-Warning `n"Full mode selected. This will last a lot more compared to the normal mode. Reboot is highly recommended" }
+if ($normal) {
+    Write-Step "Removing temporary files"
+    Write-Host `n"--> Deleting temporary files with cleanmgr..."
+    cleanmgr.exe /VERYLOWDISK
+    Write-Host "Done!"
 
-Write-Step "Removing temporary files"
-Write-Host `n"--> Deleting temporary files with cleanmgr..."
-cleanmgr.exe /VERYLOWDISK
-Write-Host "Done!"
+    Write-Host `n"--> Removing any queued updates..."
+    Dism.exe /online /Cleanup-Image /StartComponentCleanup /ResetBase | Out-Null
+    Stop-Service -Name "wuauserv", "UsoSvc", "bits", "dosvc" -ErrorAction SilentlyContinue
+    Remove-Item -Path "$env:SystemRoot\SoftwareDistribution\*" -Recurse -ErrorAction SilentlyContinue
+    Write-Host "Done!"
 
-Write-Host `n"--> Removing any queued updates..."
-Dism.exe /online /Cleanup-Image /StartComponentCleanup /ResetBase | Out-Null
-Stop-Service -Name "wuauserv", "UsoSvc", "bits", "dosvc" -ErrorAction SilentlyContinue
-Remove-Item -Path "$env:SystemRoot\SoftwareDistribution\*" -Recurse -ErrorAction SilentlyContinue
-Write-Host "Done!"
-
-Write-Host `n"--> Removing network cache..."
-arp -d * | Out-Null; nbtstat -RR | Out-Null; ipconfig /flushdns | Out-Null; ipconfig /registerdns | Out-Null
-Write-Host "Done!"
-
+    Write-Host `n"--> Removing network cache..."
+    arp -d * | Out-Null; nbtstat -RR | Out-Null; ipconfig /flushdns | Out-Null; ipconfig /registerdns | Out-Null
+    Write-Host "Done!"
+}
 if ($full) {
     Write-Step "Creating a restore point"
     Enable-ComputerRestore -Drive "$Env:SYSTEMDRIVE\"
